@@ -1,6 +1,6 @@
 {
   "name": "smapp",
-  "productName": "Spacemesh Wallet",
+  "productName": "Spacemesh",
   "version": "0.0.1",
   "author": {
     "name": "Spacemech",
@@ -19,12 +19,11 @@
     "flow-typed": "rm -rf flow-typed && flow-typed install",
     "lint": "cross-env NODE_ENV=development eslint --format=pretty .",
     "lint-ci": "npm run flow && npm run lint",
-    "package": "npm run build && electron-builder build --publish never",
-    "package-all": "npm run build && node ./configs/packageConfig.js mac windows linux",
-    "package-mac": "npm run build && node ./configs/packageConfig.js mac",
-    "package-win": "npm run build && node ./configs/packageConfig.js windows",
-    "package-linux": "npm run build && node ./configs/packageConfig.js linux",
-    "package-ci": "npm run package-all && electron-builder --publish onTag",
+    "package-all": "npm run build && node ./packagerScript.js --target mwl",
+    "package-mac": "npm run build && node ./packagerScript.js --target mac",
+    "package-win": "npm run build && node ./packagerScript.js --target windows",
+    "package-linux": "npm run build && node ./packagerScript.js --target linux",
+    "package-ci": "npm run build && node ./packagerScript.js --target mwl --publish always",
     "postinstall": "npm run flow-typed && electron-builder install-app-deps package.json && npm run build-dll",
     "prestart": "npm run build",
     "start": "cross-env NODE_ENV=production electron ./desktop/main.prod.js",
@@ -79,21 +78,21 @@
     "electron-devtools-installer": "2.2.4",
     "electron-log": "3.0.7",
     "electron-updater": "4.1.2",
-    "eslint": "6.1.0",
-    "eslint-config-airbnb": "17.1.1",
-    "eslint-config-prettier": "6.0.0",
+    "eslint": "6.2.2",
+    "eslint-config-airbnb": "18.0.1",
+    "eslint-config-prettier": "6.1.0",
     "eslint-formatter-pretty": "2.1.1",
     "eslint-import-resolver-webpack": "0.11.1",
-    "eslint-plugin-flowtype": "3.12.2",
+    "eslint-plugin-flowtype": "4.2.0",
     "eslint-plugin-import": "2.18.2",
-    "eslint-plugin-jest": "22.14.1",
+    "eslint-plugin-jest": "22.15.2",
     "eslint-plugin-jsx-a11y": "6.2.3",
     "eslint-plugin-prettier": "3.1.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-react": "7.14.3",
     "exports-loader": "0.7.0",
     "file-loader": "4.1.0",
-    "flow-bin": "0.104.0",
+    "flow-bin": "0.106.2",
     "flow-runtime": "0.17.0",
     "flow-typed": "2.6.1",
     "hoist-non-react-statics": "3.3.0",
@@ -112,6 +111,7 @@
     "auto-launch": "5.0.5",
     "bip39": "3.0.2",
     "electron-store": "4.0.0",
+    "find-process": "1.4.2",
     "grpc": "1.22.2",
     "js-xdr": "1.1.2",
     "pbkdf2": "3.0.17",
@@ -125,6 +125,6 @@
     "redux-thunk": "2.3.0",
     "source-map-support": "0.5.13",
     "styled-components": "5.0.0-alpha.2",
-    "systeminformation": "4.14.4"
+    "systeminformation": "4.14.16"
   }
 }
