@@ -1,6 +1,5 @@
 // @flow
 export { default as Loader } from './Loader';
-export { default as BaseWrapper } from './BaseWrapper';
 export { default as Button } from './Button';
 export { default as SecondaryButton } from './SecondaryButton';
 export { default as Input } from './Input';
@@ -9,5 +8,8 @@ export { default as ErrorPopup } from './ErrorPopup';
 export { default as CorneredWrapper } from './CorneredWrapper';
 export { default as StepsContainer } from './StepsContainer';
 export { default as Tooltip } from './Tooltip';
+export { default as NavTooltip } from './NavTooltip';
 export { default as WrapperWith2SideBars } from './WrapperWith2SideBars';
 export { default as DropDown } from './DropDown';
+export { default as Banner } from './Banner';
+export { default as SmallHorizontalPanel } from './SmallHorizontalPanel';
