@@ -17,8 +17,8 @@ const Wrapper = styled.div`
 
 const Header = styled.div`
   font-family: SourceCodeProBold;
-  font-size: 16px;
-  line-height: 22px;
+  font-size: 15px;
+  line-height: 20px;
   color: ${({ isStandalone }) => (isStandalone ? smColors.white : smColors.realBlack)};
 `;
 
@@ -55,22 +55,22 @@ const ButtonsWrapper = styled.div`
 `;
 
 const inputStyle1 = { margin: '10px 0 10px 10px' };
-const inputStyle2 = { margin: '10px 10px 10px 10px' };
-const inputStyle3 = { margin: '10px 0' };
-const inputStyle4 = { marginBottom: '10px' };
+const inputStyle2 = { margin: '10px 0' };
+const inputStyle3 = { marginBottom: '10px' };
 
 type Props = {
   isStandalone: boolean,
   addToContacts: Action,
   updateTransaction: Action,
   initialAddress?: string,
-  onCompleteAction: () => void
+  onCompleteAction: () => void,
+  onCancel: () => void
 };
 
 type State = {
   address: string,
+  initialAddress: string,
   nickname: string,
-  email: string,
   hasError: boolean,
   errorMsg: string
 };
@@ -80,16 +80,16 @@ class CreateNewContact extends Component<Props, State> {
     super(props);
     this.state = {
       address: props.initialAddress || '',
+      initialAddress: props.initialAddress || '',
       nickname: '',
-      email: '',
       hasError: false,
       errorMsg: ''
     };
   }
 
   render() {
-    const { isStandalone, onCompleteAction } = this.props;
-    const { address, nickname, email, hasError, errorMsg } = this.state;
+    const { isStandalone, onCancel } = this.props;
+    const { address, nickname, hasError, errorMsg } = this.state;
     return (
       <Wrapper isStandalone={isStandalone}>
         <Header isStandalone={isStandalone}>
@@ -104,40 +104,38 @@ class CreateNewContact extends Component<Props, State> {
               placeholder="Nickname"
               onChange={({ value }) => this.setState({ nickname: value, hasError: false })}
               maxLength="50"
-              style={isStandalone ? inputStyle3 : inputStyle1}
+              style={isStandalone ? inputStyle2 : inputStyle1}
             />
             <Input
               value={address}
               placeholder="Wallet address"
               onChange={({ value }) => this.setState({ address: value, hasError: false })}
               maxLength="64"
-              style={isStandalone ? inputStyle4 : inputStyle1}
-            />
-            <Input
-              value={email}
-              placeholder="Email (optional)"
-              onChange={({ value }) => this.setState({ email: value, hasError: false })}
-              maxLength="150"
-              style={isStandalone ? inputStyle4 : inputStyle2}
+              style={isStandalone ? inputStyle3 : inputStyle1}
+              onFocus={this.handleFocus}
             />
             {hasError && <ErrorPopup onClick={() => this.setState({ hasError: false })} text={errorMsg} style={{ bottom: 60, left: 'calc(50% - 90px)' }} />}
           </InputWrapperUpperPart>
           <InputWrapperLowerPart />
         </InputsWrapper>
         <ButtonsWrapper>
-          <Link onClick={onCompleteAction} text="CANCEL" style={{ color: smColors.mediumGray, marginRight: 15 }} />
-          <Link onClick={this.createContact} text="CREATE" />
+          <Link onClick={onCancel} text="CANCEL" style={{ color: smColors.disabledGray, marginRight: 15 }} />
+          <Link onClick={this.createContact} text="CREATE" style={{ color: smColors.white }} />
         </ButtonsWrapper>
       </Wrapper>
     );
   }
 
   static getDerivedStateFromProps(props: Props, prevState: State) {
-    if (props.initialAddress && props.initialAddress !== prevState.address) {
-      return { address: props.initialAddress };
+    if (props.initialAddress !== prevState.initialAddress) {
+      return { address: props.initialAddress, initialAddress: props.initialAddress };
     }
     return null;
   }
+
+  handleFocus = ({ target }: { target: Object }) => {
+    target.select();
+  };
 
   createContact = async () => {
     const { addToContacts, onCompleteAction, updateTransaction } = this.props;
@@ -145,9 +143,9 @@ class CreateNewContact extends Component<Props, State> {
     if (errorMsg) {
       this.setState({ hasError: true, errorMsg });
     } else {
-      const { address, nickname, email } = this.state;
+      const { address, nickname } = this.state;
       try {
-        await addToContacts({ contact: { address, nickname, email } });
+        await addToContacts({ contact: { address, nickname } });
         await updateTransaction({ tx: { address, nickname, isSavedContact: true }, updateAll: true });
         onCompleteAction();
       } catch (error) {
@@ -159,7 +157,7 @@ class CreateNewContact extends Component<Props, State> {
   };
 
   validate = () => {
-    const { nickname, address, email } = this.state;
+    const { nickname, address } = this.state;
     const nicknameRegex = /^([a-zA-Z0-9_-])$/;
     if (nicknameRegex.test(nickname)) {
       return 'Nickname is missing or invalid';
@@ -167,11 +165,6 @@ class CreateNewContact extends Component<Props, State> {
     const addressRegex = /\b[a-zA-Z0-9]{64}\b/;
     if (!addressRegex.test(address)) {
       return 'Address is invalid';
-    }
-    // eslint-disable-next-line no-useless-escape
-    const emailRegex = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    if (!!email && !emailRegex.test(email)) {
-      return 'Must enter a valid email';
     }
     return '';
   };
