@@ -3,23 +3,13 @@ import React, { Component } from 'react';
 import styled from 'styled-components';
 import { Tooltip, ErrorPopup } from '/basicComponents';
 import { tooltip } from '/assets/images';
-import { smColors } from '/vars';
+import { smColors, nodeConsts } from '/vars';
 
 const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
   flex: 1;
   justify-content: space-between;
-`;
-
-const HeaderWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-`;
-
-const DdWrapper = styled.div`
-  flex: 1;
-  margin-right: 100px;
 `;
 
 const SelectorUpperPart = styled.div`
@@ -37,9 +27,11 @@ const SelectorUpperPart = styled.div`
   height: 120px;
   padding: 10px;
   cursor: inherit;
-  &:hover {
-    background-color: ${({ hasError }) => (hasError ? smColors.orange : smColors.realBlack)};
-  }
+  ${({ isDisabled, hasError }) =>
+    !isDisabled &&
+    `&:hover {
+    background-color: ${hasError ? smColors.orange : smColors.realBlack};
+  }`}
   ${({ isSelected, hasError }) => (hasError ? `background-color: ${smColors.orange}` : `background-color: ${isSelected ? smColors.realBlack : smColors.darkGray}`)}
 `;
 
@@ -53,9 +45,11 @@ const SelectorLowerPart = styled.div`
   width: 160px;
   height: 120px;
   cursor: inherit;
-  &:hover {
-    border: 1px solid ${({ hasError }) => (hasError ? smColors.orange : smColors.realBlack)};
-  }
+  ${({ isDisabled, hasError }) =>
+    !isDisabled &&
+    `&:hover {
+    border: 1px solid ${hasError ? smColors.orange : smColors.realBlack};
+  }`}
   ${({ isSelected, hasError }) => (hasError ? `border: 1px solid ${smColors.orange}` : `border: 1px solid ${isSelected ? smColors.realBlack : smColors.darkGray}`)}
 `;
 
@@ -63,15 +57,17 @@ const SelectorWrapper = styled.div`
   position: relative;
   width: 165px;
   height: 125px;
-  cursor: pointer;
-  &:active ${SelectorUpperPart} {
+  ${({ isDisabled }) =>
+    !isDisabled &&
+    `cursor: pointer;
+    &:active ${SelectorUpperPart} {
     transform: translate3d(-5px, 5px, 0);
     transition: transform 0.2s cubic-bezier;
     background-color: ${smColors.black};
   }
   &:active ${SelectorLowerPart} {
     border: 1px solid ${smColors.black};
-  }
+  }`}
 `;
 
 const SelectorsWrapper = styled.div`
@@ -81,23 +77,16 @@ const SelectorsWrapper = styled.div`
 `;
 
 const TextWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  justify-content: space-between;
-  cursor: inherit;
-`;
-
-const ComplexTextWrapper = styled.div`
   position: relative;
   display: flex;
   flex-direction: row;
   align-items: center;
+  cursor: inherit;
 `;
 
 const Text = styled.div`
-  font-size: 13px;
-  line-height: 17px;
+  font-size: 20px;
+  line-height: 35px;
   color: ${smColors.white};
   cursor: inherit;
 `;
@@ -120,18 +109,26 @@ const TooltipWrapper = styled.div`
   }
 `;
 
+const ErrorPopupWrapper = styled.div`
+  position: relative;
+  width: 305px;
+  height: 65px;
+  margin: 10px auto 0;
+`;
+
 type Props = {
-  onClick: Function,
-  freeSpace: number,
-  selectedItemIndex: number
+  onClick: ({ index: number }) => void,
+  freeSpace: number
 };
 
 type State = {
+  selectedCommitmentIndex: number,
   hasInsufficientSpace: boolean
 };
 
 class CommitmentSelector extends Component<Props, State> {
   state = {
+    selectedCommitmentIndex: 0,
     hasInsufficientSpace: false
   };
 
@@ -139,35 +136,31 @@ class CommitmentSelector extends Component<Props, State> {
     const { hasInsufficientSpace } = this.state;
     return (
       <Wrapper>
-        <HeaderWrapper>
-          <DdWrapper />
-          {hasInsufficientSpace && <ErrorPopup onClick={this.closeErrorPopup} text="This partition doesn't have enough free space, please select another" />}
-        </HeaderWrapper>
         <SelectorsWrapper>{this.renderSelector()}</SelectorsWrapper>
+        <ErrorPopupWrapper>
+          {hasInsufficientSpace && <ErrorPopup onClick={this.closeErrorPopup} text="This partition doesn't have enough free space, please select another" />}
+        </ErrorPopupWrapper>
       </Wrapper>
     );
   }
 
   renderSelector = () => {
-    const { selectedItemIndex } = this.props;
-    const { hasInsufficientSpace } = this.state;
+    const { selectedCommitmentIndex, hasInsufficientSpace } = this.state;
     const selectors = [];
-    for (let i = 1; i < 4; i += 1) {
+    for (let i = 0; i < 3; i += 1) {
+      const clickHandler = i === 0 ? () => this.handleClick({ index: i }) : null;
       selectors.push(
-        <SelectorWrapper onClick={() => this.handleClick({ index: i - 1 })} key={i} style={{ zIndex: 4 - i }}>
-          <SelectorUpperPart hasError={hasInsufficientSpace && selectedItemIndex === 0} isSelected={selectedItemIndex === i - 1}>
+        <SelectorWrapper onClick={clickHandler} key={i} style={{ zIndex: 3 - i }} isDisabled={i !== 0}>
+          <SelectorUpperPart hasError={hasInsufficientSpace && selectedCommitmentIndex === i} isSelected={selectedCommitmentIndex === i} isDisabled={i !== 0}>
             <TextWrapper>
-              <ComplexTextWrapper>
-                <Text>{200 * i + 5}GB</Text>
-                <TooltipWrapper>
-                  <TooltipIcon src={tooltip} />
-                  <CustomTooltip text="The download of spacemesh requires 5GB of space in addition to the amount you choose to allocate for mining" />
-                </TooltipWrapper>
-              </ComplexTextWrapper>
-              <Text>{10 * i} SMC/WEEK</Text>
+              <Text>{nodeConsts.COMMITMENT_SIZE * (i + 1)} GB</Text>
+              <TooltipWrapper>
+                <TooltipIcon src={tooltip} />
+                <CustomTooltip text="The download of spacemesh requires 5GB of space in addition to the amount you choose to allocate for mining" />
+              </TooltipWrapper>
             </TextWrapper>
           </SelectorUpperPart>
-          <SelectorLowerPart />
+          <SelectorLowerPart hasError={hasInsufficientSpace && selectedCommitmentIndex === i} isDisabled={i !== 0} />
         </SelectorWrapper>
       );
     }
@@ -176,17 +169,15 @@ class CommitmentSelector extends Component<Props, State> {
 
   handleClick = ({ index }: { index: number }) => {
     const { freeSpace, onClick } = this.props;
-    const totalRequiredSpace = 200 * index + 5;
-    onClick({ index });
-    if (freeSpace < totalRequiredSpace) {
-      this.setState({ hasInsufficientSpace: true });
-    }
+    const totalRequiredSpace = nodeConsts.COMMITMENT_SIZE * (index + 1);
+    onClick({ index: freeSpace < totalRequiredSpace ? 0 : totalRequiredSpace });
+    this.setState({ selectedCommitmentIndex: index, hasInsufficientSpace: freeSpace < totalRequiredSpace });
   };
 
   closeErrorPopup = () => {
     const { onClick } = this.props;
     onClick({ index: -1 });
-    this.setState({ hasInsufficientSpace: false });
+    this.setState({ selectedCommitmentIndex: -1, hasInsufficientSpace: false });
   };
 }
 
