@@ -1,4 +1,4 @@
 export { default as SettingsSection } from './SettingsSection';
 export { default as SettingRow } from './SettingRow';
-export { default as ChangePassphrase } from './ChangePassphrase';
+export { default as ChangePassword } from './ChangePassword';
 export { default as SideMenu } from './SideMenu';
