@@ -71,8 +71,8 @@ class TransactionsMeta extends PureComponent<Props> {
     const totalCoinsMeta = [{ title: 'TOTAL MINED', coins: totalMined }, { title: 'TOTAL SENT', coins: totalSent }, { title: 'TOTAL RECEIVED', coins: totalReceived }];
 
     return (
-      <React.Fragment>
-        <BoldText>traffic</BoldText>
+      <>
+        <BoldText>activity</BoldText>
         <BoldText>--</BoldText>
         <Text style={{ marginBottom: 27 }}>{`${filterName.replace(/^\w/, (c) => c.toUpperCase())} coins: ${totalFilteredCoins}`}</Text>
         {coinsMeta.map((coinMeta) => (
@@ -92,7 +92,7 @@ class TransactionsMeta extends PureComponent<Props> {
             <Text>{`${totalMeta.coins} SMC`}</Text>
           </TextRow>
         ))}
-      </React.Fragment>
+      </>
     );
   }
 }
