@@ -35,6 +35,8 @@ const AccountName = styled.div`
 `;
 
 const Address = styled.div`
+  display: flex;
+  flex-direction: row;
   font-size: 16px;
   line-height: 22px;
   color: ${smColors.black};
@@ -42,16 +44,17 @@ const Address = styled.div`
 `;
 
 const CopyIcon = styled.img`
+  align-self: flex-end;
   width: 16px;
   height: 15px;
-  margin: 0 10px;
+  margin: 6px;
   cursor: pointer;
-`;
-
-const CopiedText = styled.div`
-  font-size: 16px;
-  line-height: 22px;
-  color: ${smColors.green};
+  &:hover {
+    opacity: 0.5;
+  }
+  &:active {
+    transform: translate3d(2px, 2px, 0);
+  }
 `;
 
 const Footer = styled.div`
@@ -86,7 +89,17 @@ const SmcText = styled.div`
   color: ${smColors.green};
 `;
 
+const CopiedText = styled.div`
+  text-align: left;
+  font-size: 16px;
+  line-height: 20px;
+  height: 20px;
+  margin-left: 6px;
+  color: ${smColors.green};
+`;
+
 type Props = {
+  walletName: string,
   accounts: Account[],
   currentAccountIndex: number,
   switchAccount: ({ index: number }) => void
@@ -97,18 +110,21 @@ type State = {
 };
 
 class AccountsOverview extends Component<Props, State> {
-  copiedTimeout: any;
+  copiedTimeout: TimeoutID;
 
   state = {
     isCopied: false
   };
 
   render() {
-    const { accounts, currentAccountIndex, switchAccount } = this.props;
+    const { walletName, accounts, currentAccountIndex, switchAccount } = this.props;
     const { isCopied } = this.state;
+    if (!accounts || !accounts.length) {
+      return null;
+    }
     const { displayName, pk, balance } = accounts[currentAccountIndex];
     return (
-      <WrapperWith2SideBars width={300} height={480} header="WALLET">
+      <WrapperWith2SideBars width={300} height={480} header={walletName}>
         <AccountDetails>
           {accounts.length > 1 ? (
             <DropDown
@@ -119,11 +135,10 @@ class AccountsOverview extends Component<Props, State> {
               rowHeight={55}
             />
           ) : (
-            this.renderAccountRow({ displayName, isCopied, pk })
+            this.renderAccountRow({ displayName, pk })
           )}
-          <CopyIcon src={copyToClipboard} />
         </AccountDetails>
-        <CopiedText>{isCopied ? 'Address had been copied to clipboard!' : ''}</CopiedText>
+        {isCopied && <CopiedText>COPIED</CopiedText>}
         <Footer>
           <BalanceHeader>BALANCE</BalanceHeader>
           <BalanceWrapper>
@@ -136,17 +151,20 @@ class AccountsOverview extends Component<Props, State> {
   }
 
   renderAccountRow = ({ displayName, pk, isInDropDown }: { displayName: string, pk: string, isInDropDown?: boolean }) => (
-    <AccountWrapper key={pk} isInDropDown={isInDropDown}>
+    <AccountWrapper isInDropDown={isInDropDown}>
       <AccountName>{displayName}</AccountName>
-      <Address>{getAbbreviatedText(pk, 6)}</Address>
+      <Address>
+        {getAbbreviatedText(pk)}
+        <CopyIcon src={copyToClipboard} onClick={this.copyPublicAddress} />
+      </Address>
     </AccountWrapper>
   );
 
   copyPublicAddress = () => {
     const { accounts, currentAccountIndex } = this.props;
     clearTimeout(this.copiedTimeout);
-    clipboard.writeText(accounts[currentAccountIndex].pk);
-    this.copiedTimeout = setTimeout(() => this.setState({ isCopied: false }), 3000);
+    clipboard.writeText(`0x${accounts[currentAccountIndex].pk}`);
+    this.copiedTimeout = setTimeout(() => this.setState({ isCopied: false }), 10000);
     this.setState({ isCopied: true });
   };
 }
