@@ -1,7 +1,8 @@
 // @flow
 import { httpService } from '/infra/httpService';
 import { createError } from '/infra/utils';
-import { Action, Dispatch } from '/types';
+import { nodeConsts } from '/vars';
+import { Action, Dispatch, GetState } from '/types';
 
 export const CHECK_NODE_CONNECTION: string = 'CHECK_NODE_CONNECTION';
 
@@ -9,8 +10,7 @@ export const SET_MINING_STATUS: string = 'SET_MINING_STATUS';
 export const INIT_MINING: string = 'INIT_MINING';
 
 export const SET_GENESIS_TIME: string = 'SET_GENESIS_TIME';
-export const SET_TOTAL_AWARDS: string = 'SET_TOTAL_AWARDS';
-export const SET_UPCOMING_REWARDS: string = 'SET_UPCOMING_REWARDS';
+export const SET_UPCOMING_AWARDS: string = 'SET_UPCOMING_AWARDS';
 
 export const SET_NODE_IP: string = 'SET_NODE_IP';
 export const SET_AWARDS_ADDRESS: string = 'SET_AWARDS_ADDRESS';
@@ -19,9 +19,10 @@ export const checkNodeConnection = (): Action => async (dispatch: Dispatch): Dis
   try {
     await httpService.checkNodeConnection();
     dispatch({ type: CHECK_NODE_CONNECTION, payload: { isConnected: true } });
+    return true;
   } catch (err) {
     dispatch({ type: CHECK_NODE_CONNECTION, payload: { isConnected: false } });
-    throw err;
+    return false;
   }
 };
 
@@ -34,30 +35,33 @@ export const getMiningStatus = (): Action => async (dispatch: Dispatch): Dispatc
   }
 };
 
-export const initMining = ({ capacity, drive, address }: { capacity: { id: number, label: string }, drive: { mountPoint: string }, address: string }): Action => async (
+export const initMining = ({ logicalDrive, commitmentSize, address }: { logicalDrive: string, commitmentSize: number, address: string }): Action => async (
   dispatch: Dispatch
 ): Dispatch => {
   try {
-    await httpService.initMining({ logicalDrive: drive.mountPoint, commitmentSize: capacity.id, address });
+    await httpService.initMining({ logicalDrive, commitmentSize, coinbase: address });
     dispatch({ type: INIT_MINING, payload: { address } });
   } catch (err) {
-    throw createError('Error setting node storage', () => initMining({ capacity, drive, address }));
+    throw createError('Error initiating mining', () => initMining({ logicalDrive, commitmentSize, address }));
   }
 };
 
 export const getGenesisTime = (): Action => async (dispatch: Dispatch): Dispatch => {
   try {
     const genesisTime = await httpService.getGenesisTime();
-    dispatch({ type: SET_GENESIS_TIME, payload: { genesisTime } });
+    dispatch({ type: SET_GENESIS_TIME, payload: { genesisTime: new Date(genesisTime).getTime() } });
   } catch (err) {
     console.error(err); // eslint-disable-line no-console
   }
 };
 
-export const getUpcomingRewards = (): Action => async (dispatch: Dispatch): Dispatch => {
+export const getUpcomingAwards = (): Action => async (dispatch: Dispatch, getState: GetState): Dispatch => {
   try {
-    const timeTillNextReward = await httpService.getUpcomingRewards();
-    dispatch({ type: SET_UPCOMING_REWARDS, payload: { timeTillNextReward } });
+    const awardLayerNumbers = await httpService.getUpcomingAwards();
+    const { genesisTime } = getState().node;
+    const currentLayer = Math.floor((new Date().getTime() - genesisTime) / nodeConsts.TIME_BETWEEN_LAYERS);
+    const futureAwardLayerNumbers = awardLayerNumbers.filter((layer) => layer >= currentLayer);
+    dispatch({ type: SET_UPCOMING_AWARDS, payload: { timeTillNextAward: nodeConsts.TIME_BETWEEN_LAYERS * (futureAwardLayerNumbers[0] - currentLayer) } });
   } catch (err) {
     console.error(err); // eslint-disable-line no-console
   }
@@ -68,7 +72,7 @@ export const setNodeIpAddress = ({ nodeIpAddress }: { nodeIpAddress: string }): 
     await httpService.setNodeIpAddress({ nodeIpAddress });
     dispatch({ type: SET_NODE_IP, payload: { nodeIpAddress } });
   } catch (err) {
-    throw err;
+    throw createError('Error setting node IP address', () => setNodeIpAddress({ nodeIpAddress }));
   }
 };
 
