@@ -1,23 +1,13 @@
 // @flow
 import { shell } from 'electron';
 import React, { PureComponent } from 'react';
-import styled from 'styled-components';
 import { CorneredContainer } from '/components/common';
-import { Button, Link, SecondaryButton } from '/basicComponents';
-import { smallHorizontalSideBar, chevronLeftWhite } from '/assets/images';
+import { Button, Link, SecondaryButton, SmallHorizontalPanel } from '/basicComponents';
+import { chevronLeftWhite } from '/assets/images';
 import type { RouterHistory } from 'react-router-dom';
 
-const SideBar = styled.img`
-  position: absolute;
-  top: -30px;
-  right: 0;
-  width: 55px;
-  height: 15px;
-`;
-
-const Buttons = styled.div`
-  margin-top: 30px;
-`;
+const secondaryBtnStyle = { position: 'absolute', bottom: 0, left: -35 };
+const btnStyle = { margin: '30px 0 15px' };
 
 type Props = {
   history: RouterHistory
@@ -27,13 +17,11 @@ class RestoreWallet extends PureComponent<Props> {
   render() {
     const { history } = this.props;
     return (
-      <CorneredContainer width={650} height={400} header="RESTORE EXISTING WALLET" subHeader="choose how you&#39;d like to restore an existing wallet">
-        <SideBar src={smallHorizontalSideBar} />
-        <SecondaryButton onClick={history.goBack} img={chevronLeftWhite} imgWidth={10} imgHeight={15} style={{ position: 'absolute', bottom: 0, left: -35 }} />
-        <Buttons>
-          <Button text="Restore From File" isPrimary={false} onClick={() => history.push('/auth/file-restore')} width={250} style={{ marginBottom: 15 }} />
-          <Button text="Restore With 12 Words" isPrimary={false} onClick={() => history.push('/auth/words-restore')} width={250} />
-        </Buttons>
+      <CorneredContainer width={650} height={400} header="RESTORE EXISTING WALLET" subHeader="Choose how you&#39;d like to restore an existing wallet">
+        <SmallHorizontalPanel />
+        <SecondaryButton onClick={history.goBack} img={chevronLeftWhite} imgWidth={10} imgHeight={15} style={secondaryBtnStyle} />
+        <Button text="RESTORE FROM FILE" isPrimary={false} onClick={() => history.push('/auth/file-restore')} width={250} style={btnStyle} />
+        <Button text="RESTORE WITH 12 WORDS" isPrimary={false} onClick={() => history.push('/auth/words-restore')} width={250} />
         <Link onClick={this.navigateToWalletGuide} text="WALLET GUIDE" style={{ marginTop: 'auto', marginRight: 'auto' }} />
       </CorneredContainer>
     );
