@@ -33,7 +33,7 @@ const SubHeader = styled.div`
 `;
 
 const SubHeaderText = styled.div`
-  font-size: 16px;
+  font-size: 15px;
   line-height: 20px;
   color: ${smColors.realBlack};
 `;
@@ -204,7 +204,7 @@ type State = {
 };
 
 class Contacts extends Component<Props, State> {
-  newContactCreatedTimeOut: any;
+  newContactCreatedTimeOut: TimeoutID;
 
   state = {
     addressToAdd: '',
@@ -217,9 +217,9 @@ class Contacts extends Component<Props, State> {
 
   render() {
     const { contacts } = this.props;
-    const { tmpSearchTerm, selectedSorting } = this.state;
+    const { tmpSearchTerm, searchTerm, selectedSorting } = this.state;
     return (
-      <WrapperWith2SideBars width={1000} height={600} header="MY CONTACTS">
+      <WrapperWith2SideBars width={1000} height={600} header="CONTACTS">
         <SearchWrapper>
           <SearchIcon src={searchIcon} />
           <Input
@@ -242,7 +242,9 @@ class Contacts extends Component<Props, State> {
             style={{ flex: '0 0 150px', borderBottom: '1px solid' }}
           />
         </ContactsSubHeader>
-        <ContactsList>{contacts && contacts.length && this.renderContacts()}</ContactsList>
+        <ContactsList>
+          {contacts && contacts.length ? this.renderContacts() : <ContactText>{searchTerm ? 'No contacts matching criteria' : 'No contacts added yet'}</ContactText>}
+        </ContactsList>
       </WrapperWith2SideBars>
     );
   }
@@ -255,7 +257,7 @@ class Contacts extends Component<Props, State> {
     const { contacts } = this.props;
     const { addressToAdd, shouldShowCreateNewContactModal, isNewContactCreated } = this.state;
     if (shouldShowCreateNewContactModal) {
-      return <CreateNewContact initialAddress={addressToAdd} onCompleteAction={this.createdNewContact} />;
+      return <CreateNewContact initialAddress={addressToAdd} onCompleteAction={this.createdNewContact} onCancel={this.cancelCreateNewContact} />;
     }
     if (isNewContactCreated) {
       return <CreatedNewContact contact={contacts[0]} action={() => this.navigateToSendCoins({ contact: contacts[0] })} />;
@@ -320,8 +322,7 @@ class Contacts extends Component<Props, State> {
     return sortedContacts.map((contact) => (
       <ContactRow key={`${contact.nickname}_${contact.address}`} onClick={() => this.navigateToSendCoins({ contact })}>
         <ContactText>{contact.nickname || 'UNKNOWN ADDRESS'}</ContactText>
-        <ContactText>{getAbbreviatedText(contact.address, 8)}</ContactText>
-        <ContactText>{contact.email}</ContactText>
+        <ContactText>{getAbbreviatedText(contact.address)}</ContactText>
         {!contact.nickname && <CreateNewContactImg onClick={() => this.setState({ addressToAdd: contact.address, shouldShowCreateNewContactModal: true })} src={addContact} />}
       </ContactRow>
     ));
@@ -331,8 +332,7 @@ class Contacts extends Component<Props, State> {
     const { searchTerm } = this.state;
     const nicknameMatch = contact.nickname && contact.nickname.toLowerCase().includes(searchTerm);
     const addressMatch = contact.address && contact.address.toLowerCase().includes(searchTerm);
-    const emailMatch = contact.email && contact.email.includes(searchTerm);
-    return nicknameMatch || addressMatch || emailMatch;
+    return nicknameMatch || addressMatch;
   };
 
   sortContacts = (c1: Contact, c2: Contact) => {
@@ -355,6 +355,10 @@ class Contacts extends Component<Props, State> {
   createdNewContact = () => {
     this.setState({ addressToAdd: '', shouldShowCreateNewContactModal: false, isNewContactCreated: true });
     this.newContactCreatedTimeOut = setTimeout(() => this.setState({ isNewContactCreated: false }), 10000);
+  };
+
+  cancelCreateNewContact = () => {
+    this.setState({ addressToAdd: '', shouldShowCreateNewContactModal: false, isNewContactCreated: false });
   };
 
   navigateToSendCoins = ({ contact }: { contact: Contact }) => {
