@@ -3,14 +3,14 @@ import { shell } from 'electron';
 import React, { Component } from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
-import { getUpcomingRewards } from '/redux/node/actions';
+import { getUpcomingAwards } from '/redux/node/actions';
 import { CorneredContainer } from '/components/common';
 import { WrapperWith2SideBars, Link, Button } from '/basicComponents';
 import { ScreenErrorBoundary } from '/components/errorHandler';
 import { playIcon, pauseIcon } from '/assets/images';
 import { smColors, nodeConsts } from '/vars';
 import type { RouterHistory } from 'react-router-dom';
-// import type { Action } from '/types';
+import type { Action } from '/types';
 
 const Wrapper = styled.div`
   display: flex;
@@ -38,7 +38,7 @@ const LogText = styled.div`
   color: ${smColors.black};
 `;
 
-const RewardText = styled(LogText)`
+const AwardText = styled(LogText)`
   color: ${smColors.green};
 `;
 
@@ -106,9 +106,9 @@ const Dots = styled(LeftText)`
 type Props = {
   isConnected: boolean,
   miningStatus: number,
-  timeTillNextReward: string,
-  // getUpcomingRewards: Action,
+  timeTillNextAward: number,
   totalEarnings: number,
+  getUpcomingAwards: Action,
   history: RouterHistory,
   location: { state?: { showIntro?: boolean } }
 };
@@ -119,6 +119,8 @@ type State = {
 };
 
 class Node extends Component<Props, State> {
+  getUpcomingAwardsInterval: IntervalID;
+
   constructor(props) {
     super(props);
     const { location } = props;
@@ -131,10 +133,10 @@ class Node extends Component<Props, State> {
   render() {
     return (
       <Wrapper>
-        <WrapperWith2SideBars width={650} height={480} header="SPACEMESH MINER" style={{ marginRight: 10 }}>
+        <WrapperWith2SideBars width={650} height={480} header="SPACEMESH FULL NODE" style={{ marginRight: 10 }}>
           {this.renderMainSection()}
         </WrapperWith2SideBars>
-        <CorneredContainer width={250} height={480} header="MINER LOG">
+        <CorneredContainer width={250} height={480} header="FULL NODE LOG">
           <LogInnerWrapper>
             <LogEntry>
               <LogText>12.09.19 - 13:00</LogText>
@@ -143,7 +145,7 @@ class Node extends Component<Props, State> {
             <LogEntrySeparator>...</LogEntrySeparator>
             <LogEntry>
               <LogText>12.09.19 - 13:10</LogText>
-              <RewardText>Network reward: 2SMC</RewardText>
+              <AwardText>Network award: 2SMC</AwardText>
             </LogEntry>
             <LogEntrySeparator>...</LogEntrySeparator>
             <LogEntry>
@@ -158,14 +160,15 @@ class Node extends Component<Props, State> {
   }
 
   async componentDidMount() {
-    // const { getUpcomingRewards } = this.props;
-    try {
-      // await getUpcomingRewards();
-    } catch (error) {
-      this.setState(() => {
-        throw error;
-      });
+    const { isConnected, miningStatus, getUpcomingAwards } = this.props;
+    if (isConnected && miningStatus === nodeConsts.IS_MINING) {
+      await getUpcomingAwards();
+      this.getUpcomingAwardsInterval = setInterval(getUpcomingAwards, nodeConsts.TIME_BETWEEN_LAYERS);
     }
+  }
+
+  componentWillUnmount(): * {
+    this.getUpcomingAwardsInterval && clearInterval(this.getUpcomingAwardsInterval);
   }
 
   renderMainSection = () => {
@@ -182,12 +185,16 @@ class Node extends Component<Props, State> {
   renderIntro = () => {
     return [
       <BoldText key="1">Success! You are now a Spacemesh testnet member!</BoldText>,
-      <Text key="2">* You will receive a desktop notification about your mining rewards</Text>,
+      <Text key="2">* You will receive a desktop notification about your mining awards</Text>,
       <Text key="3">* You can close this app, Mining still happens in the background</Text>,
       <BoldText key="4">Important:</BoldText>,
       <Text key="5">* Leave your computer on 24/7 to mine</Text>,
       <Text key="6">* Disable your computer from going to sleep</Text>,
-      <Text key="7" style={{ display: 'flex', flexDirection: 'row' }}>
+      <Text key="7">
+        * Important: configure your network to accept incoming app connections.
+        <Link onClick={this.navigateToNetConfigGuide} text="Learn more." style={{ display: 'inline', fontSize: '16px', lineHeight: '20px' }} />
+      </Text>,
+      <Text key="8" style={{ display: 'flex', flexDirection: 'row' }}>
         *&nbsp;
         <Link onClick={this.navigateToMiningGuide} text="Learn more about Spacemesh Mining" style={{ fontSize: '16px', lineHeight: '20px' }} />
       </Text>,
@@ -215,19 +222,19 @@ class Node extends Component<Props, State> {
   };
 
   renderNodeDashboard = () => {
-    const { isConnected, timeTillNextReward, totalEarnings } = this.props;
+    const { isConnected, timeTillNextAward, totalEarnings } = this.props;
     const { isMiningPaused } = this.state;
     return [
       <Status key="status" isConnected={isConnected}>
         {isConnected ? 'Connected!' : 'Not connected!'}
       </Status>,
       <TextWrapper key="1">
-        <LeftText>Upcoming reward in</LeftText>
+        <LeftText>Upcoming award in</LeftText>
         <Dots>....................................</Dots>
-        <RightText>{timeTillNextReward}</RightText>
+        <RightText>{Math.floor(timeTillNextAward / 1000)} min</RightText>
       </TextWrapper>,
       <TextWrapper key="2">
-        <LeftText>Total Rewards</LeftText>
+        <LeftText>Total Awards</LeftText>
         <Dots>....................................</Dots>
         <GreenText>{totalEarnings} SMC</GreenText>
       </TextWrapper>,
@@ -248,17 +255,19 @@ class Node extends Component<Props, State> {
   pauseResumeMining = () => {};
 
   navigateToMiningGuide = () => shell.openExternal('https://testnet.spacemesh.io/#/guide/setup');
+
+  navigateToNetConfigGuide = () => shell.openExternal('https://testnet.spacemesh.io/#/netconfig');
 }
 
 const mapStateToProps = (state) => ({
   isConnected: state.node.isConnected,
   miningStatus: state.node.miningStatus,
-  timeTillNextReward: state.node.timeTillNextReward,
+  timeTillNextAward: state.node.timeTillNextAward,
   totalEarnings: state.node.totalEarnings
 });
 
 const mapDispatchToProps = {
-  getUpcomingRewards
+  getUpcomingAwards
 };
 
 Node = connect<any, any, _, _, _, _>(
