@@ -6,11 +6,11 @@ import { connect } from 'react-redux';
 import { initMining } from '/redux/node/actions';
 import { CorneredContainer } from '/components/common';
 import { ScreenErrorBoundary } from '/components/errorHandler';
-import { StepsContainer, Button, SecondaryButton, Link } from '/basicComponents';
+import { StepsContainer, Button, SecondaryButton, Link, SmallHorizontalPanel } from '/basicComponents';
 import { Carousel, CommitmentSelector } from '/components/node';
 import { diskStorageService } from '/infra/diskStorageService';
-import { smallHorizontalSideBar, chevronLeftWhite } from '/assets/images';
-import { smColors } from '/vars';
+import { chevronLeftWhite } from '/assets/images';
+import { smColors, nodeConsts } from '/vars';
 import type { RouterHistory } from 'react-router-dom';
 import type { Account, Action } from '/types';
 
@@ -19,15 +19,14 @@ const Wrapper = styled.div`
   flex-direction: row;
 `;
 
-const SideBar = styled.img`
-  position: absolute;
-  top: -30px;
-  right: 0;
-  width: 55px;
-  height: 15px;
+const SubHeader = styled.div`
+  margin-bottom: 20px;
+  font-size: 16px;
+  line-height: 20px;
+  color: ${smColors.black};
 `;
 
-const BottomPart = styled.div`
+const Footer = styled.div`
   display: flex;
   flex: 1;
   flex-direction: row;
@@ -39,18 +38,32 @@ const DriveName = styled.span`
   color: ${smColors.darkerGreen};
 `;
 
+const EmptyState = styled.div`
+  display: flex;
+  flex-direction: row;
+`;
+
+const Text = styled.div`
+  font-size: 14px;
+  line-height: 18px;
+  color: ${smColors.black};
+`;
+
+const bntStyle = { position: 'absolute', bottom: 0, left: -35 };
+
 type Props = {
   accounts: Account[],
   initMining: Action,
+  isConnected: boolean,
   history: RouterHistory,
   location: { state?: { isOnlyNodeSetup: boolean } }
 };
 
 type State = {
-  drives: { id: string, label: string, availableDiskSpace: { bytes: number, readable: string }, capacityAllocationsList: { id: number, label: string }[] }[],
+  drives: { id: string, label: string, mountPoint: string, availableDiskSpace: string }[],
   subMode: 2 | 3,
   selectedDriveIndex: number,
-  selectedCapacityIndex: number
+  selectedCommitmentSize: number
 };
 
 class NodeSetup extends Component<Props, State> {
@@ -58,80 +71,102 @@ class NodeSetup extends Component<Props, State> {
 
   steps: Array<string>;
 
+  header: string;
+
   constructor(props: Props) {
     super(props);
     const { location } = props;
     this.isOnlyNodeSetup = !!location?.state?.isOnlyNodeSetup;
+    this.header = this.isOnlyNodeSetup ? 'SETUP NODE' : 'SETUP WALLET + FULL NODE';
     this.steps = ['SELECT DRIVE', 'ALLOCATE SPACE'];
     if (!this.isOnlyNodeSetup) {
-      this.steps = ['SETUP WALLET + MINER', 'PROTECT WALLET'].concat(this.steps);
+      this.steps = ['PROTECT WALLET'].concat(this.steps);
     }
     this.state = {
       drives: [],
       subMode: 2,
       selectedDriveIndex: -1,
-      selectedCapacityIndex: -1
+      selectedCommitmentSize: 0
     };
   }
 
   render() {
-    const { drives, subMode, selectedDriveIndex, selectedCapacityIndex } = this.state;
+    const { isConnected } = this.props;
+    const { subMode, selectedDriveIndex, selectedCommitmentSize } = this.state;
+    const adjustedSubStep = this.isOnlyNodeSetup ? subMode % 2 : subMode - 1;
     return (
       <Wrapper>
-        <StepsContainer steps={this.steps} currentStep={this.isOnlyNodeSetup ? subMode % 2 : subMode} />
-        <CorneredContainer width={650} height={400} header={this.steps[subMode]} subHeader={this.renderSubHeader(subMode)}>
-          <SideBar src={smallHorizontalSideBar} />
-          <SecondaryButton onClick={this.handleBackBtn} img={chevronLeftWhite} imgWidth={10} imgHeight={15} style={{ position: 'absolute', bottom: 0, left: -35 }} />
-          {subMode === 2 && drives.length && (
-            <Carousel data={drives} selectedItemIndex={selectedDriveIndex} onClick={({ index }) => this.setState({ selectedDriveIndex: index })} />
-          )}
-          {subMode === 3 && (
-            <CommitmentSelector
-              freeSpace={drives[selectedDriveIndex].availableDiskSpace.readable.split(' ')[0]}
-              onClick={({ index }) => this.setState({ selectedCapacityIndex: index })}
-              selectedItemIndex={selectedCapacityIndex}
-            />
-          )}
-          <BottomPart>
+        <StepsContainer header={this.header} steps={this.steps} currentStep={adjustedSubStep} />
+        <CorneredContainer width={650} height={400} header={this.steps[adjustedSubStep]}>
+          <SmallHorizontalPanel />
+          <SecondaryButton onClick={this.handleBackBtn} img={chevronLeftWhite} imgWidth={10} imgHeight={15} style={bntStyle} />
+          {this.renderSubMode()}
+          <Footer>
             <Link onClick={this.navigateToExplanation} text="SETUP GUIDE" />
-            <Button onClick={this.nextAction} text="NEXT" isDisabled={(subMode === 2 && selectedDriveIndex === -1) || (subMode === 3 && selectedCapacityIndex === -1)} />
-          </BottomPart>
+            <Button
+              onClick={this.nextAction}
+              text="NEXT"
+              isDisabled={(subMode === 2 && selectedDriveIndex === -1) || ((subMode === 3 && selectedCommitmentSize === 0) || !isConnected)}
+            />
+          </Footer>
         </CorneredContainer>
       </Wrapper>
     );
   }
 
-  componentDidMount() {
-    this.getDrivesList();
+  async componentDidMount() {
+    const drives = await diskStorageService.getDriveList();
+    const selectedDriveIndex = drives.length ? 0 : -1;
+    const selectedCommitmentSize = drives.length ? nodeConsts.COMMITMENT_SIZE : 0;
+    this.setState({ drives, selectedDriveIndex, selectedCommitmentSize });
   }
 
-  renderSubHeader = (subMode: number) => {
-    const { drives, selectedDriveIndex } = this.state;
-    return subMode === 2 ? (
-      <span>
-        select the hard drive you&#39;d like to use for mining
-        <br />
-        You will need at least 160 GB free space to setup miner
-      </span>
-    ) : (
-      <div>
-        Allocate how much space on <DriveName>{drives[selectedDriveIndex].label} HARD DRIVE</DriveName> you would
-        <br />
-        like the mining node to use
-      </div>
+  renderSubMode = () => {
+    const { subMode, drives, selectedDriveIndex } = this.state;
+    if (subMode === 2) {
+      return (
+        <>
+          <SubHeader>
+            --
+            <br />
+            Select the hard drive you&#39;d like to use for mining
+            <br />
+            {`You will need at least ${nodeConsts.COMMITMENT_SIZE} GB free space to setup full node`}
+          </SubHeader>
+          {drives.length ? (
+            <Carousel data={drives} onClick={({ index }) => this.setState({ selectedDriveIndex: index })} />
+          ) : (
+            <EmptyState>
+              <Text>Insufficient disk space. You need a local hard drive with at least 256GB of free space to setup mining.</Text>
+              <Link onClick={this.navigateToNodeSetupGuide} text="Learn more..." />
+            </EmptyState>
+          )}
+        </>
+      );
+    }
+    return (
+      <>
+        <SubHeader>
+          --
+          <br />
+          Allocate how much space on <DriveName>{drives[selectedDriveIndex].label}</DriveName> you would
+          <br />
+          like the mining node to use
+        </SubHeader>
+        <CommitmentSelector freeSpace={drives[selectedDriveIndex].availableDiskSpace} onClick={({ index }) => this.setState({ selectedCommitmentSize: index })} />
+      </>
     );
-  };
-
-  getDrivesList = async () => {
-    const drives = await diskStorageService.getDriveList();
-    this.setState({ drives });
   };
 
   setupAndInitMining = async () => {
     const { initMining, accounts, history } = this.props;
-    const { drives, selectedCapacityIndex, selectedDriveIndex } = this.state;
+    const { drives, selectedCommitmentSize, selectedDriveIndex } = this.state;
     try {
-      await initMining({ capacity: drives[selectedDriveIndex].capacityAllocationsList[selectedCapacityIndex], drive: drives[selectedDriveIndex], address: accounts[0].pk });
+      await initMining({
+        logicalDrive: drives[selectedDriveIndex].mountPoint,
+        commitmentSize: selectedCommitmentSize * 1073741824,
+        address: accounts[0].pk
+      });
       history.push('/main/node', { showIntro: true });
     } catch (error) {
       this.setState(() => {
@@ -160,9 +195,12 @@ class NodeSetup extends Component<Props, State> {
   };
 
   navigateToExplanation = () => shell.openExternal('https://testnet.spacemesh.io/#/guide/setup');
+
+  navigateToNodeSetupGuide = () => shell.openExternal('https://testnet.spacemesh.io/#/guide/setup?id=step-2-full-node-amp-mining-setup');
 }
 
 const mapStateToProps = (state) => ({
+  isConnected: state.node.isConnected,
   accounts: state.wallet.accounts
 });
 
@@ -175,5 +213,5 @@ NodeSetup = connect<any, any, _, _, _, _>(
   mapDispatchToProps
 )(NodeSetup);
 
-NodeSetup = ScreenErrorBoundary(NodeSetup, true);
+NodeSetup = ScreenErrorBoundary(NodeSetup);
 export default NodeSetup;
