@@ -1,5 +1,6 @@
 // @flow
 import React, { Component } from 'react';
+import { shell } from 'electron';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
 import type { RouterHistory } from 'react-router-dom';
@@ -64,6 +65,7 @@ const TransactionsListWrapper = styled.div`
   flex: 1;
   overflow-x: hidden;
   overflow-y: scroll;
+  margin-bottom: 6px;
 `;
 
 const RightPaneWrapper = styled(CorneredWrapper)`
@@ -129,7 +131,6 @@ class Transactions extends Component<Props, State> {
         <SecondaryButton onClick={history.goBack} img={chevronLeftWhite} imgWidth={7} imgHeight={10} style={{ position: 'absolute', left: -35, bottom: 0 }} />
         <WrapperWith2SideBars width={680} height={480} header="TRANSACTION LOG" style={{ marginRight: 10 }}>
           <BoldText>Latest transactions</BoldText>
-          <Link onClick={this.navigateToContacts} text="MY CONTACTS" style={{ position: 'absolute', top: 30, left: 550 }} />
           <TransactionsListWrapper>
             {filteredTransactions && filteredTransactions.length ? (
               filteredTransactions.map((tx, index) => (
@@ -139,9 +140,10 @@ class Transactions extends Component<Props, State> {
               <Text>No transactions here yet</Text>
             )}
           </TransactionsListWrapper>
+          <Link onClick={this.navigateToGuide} text="TRANSACTIONS GUIDE" />
         </WrapperWith2SideBars>
         {addressToAdd ? (
-          <CreateNewContact isStandalone initialAddress={addressToAdd} onCompleteAction={this.handleCompleteAction} />
+          <CreateNewContact isStandalone initialAddress={addressToAdd} onCompleteAction={this.handleCompleteAction} onCancel={this.cancelCreatingNewContact} />
         ) : (
           <RightPaneWrapper>
             <DropDown
@@ -185,12 +187,16 @@ class Transactions extends Component<Props, State> {
     });
   };
 
+  cancelCreatingNewContact = () => {
+    this.setState({ addressToAdd: '' });
+  };
+
   filterTransactions = ({ index, transactions }) => {
     const oneDayInMs = 86400000;
     const spanInDays = [1, 30, 365];
     return transactions.data.filter((transaction: Tx) => {
       const startDate = +new Date() - spanInDays[index] * oneDayInMs;
-      return transaction.date >= startDate;
+      return +new Date(transaction.date) >= startDate;
     });
   };
 
@@ -198,6 +204,8 @@ class Transactions extends Component<Props, State> {
     const { history } = this.props;
     history.push('/main/contacts');
   };
+
+  navigateToGuide = () => shell.openExternal('https://testnet.spacemesh.io/#/wallet');
 }
 
 const mapStateToProps = (state) => ({
