@@ -7,10 +7,10 @@ import { getBalance, setCurrentAccount } from '/redux/wallet/actions';
 import routes from '/routes';
 import { AccountsOverview } from '/components/wallet';
 import { ScreenErrorBoundary } from '/components/errorHandler';
-import { CorneredWrapper } from '/basicComponents';
+import { CorneredWrapper, SmallHorizontalPanel } from '/basicComponents';
 import { localStorageService } from '/infra/storageService';
 import smColors from '/vars/colors';
-import { backup, leftSideTIcon, smallHorizontalSideBar } from '/assets/images';
+import { backup, leftSideTIcon } from '/assets/images';
 import type { Account, Action } from '/types';
 import type { RouterHistory } from 'react-router-dom';
 
@@ -67,19 +67,13 @@ const RightSection = styled.div`
   height: 100%;
 `;
 
-const HorizontalBar = styled.img`
-  position: absolute;
-  top: -25px;
-  right: 0;
-  width: 70px;
-  height: 15px;
-`;
-
 type Props = {
+  displayName: string,
   accounts: Account[],
   currentAccountIndex: number,
   getBalance: Action,
   setCurrentAccount: Action,
+  isConnected: boolean,
   history: RouterHistory
 };
 
@@ -91,13 +85,15 @@ type State = {
 };
 
 class Wallet extends Component<Props, State> {
+  getBalanceInterval: IntervalID;
+
   render() {
-    const { accounts, currentAccountIndex, setCurrentAccount } = this.props;
+    const { displayName, accounts, currentAccountIndex, setCurrentAccount } = this.props;
     const hasBackup = !!localStorageService.get('hasBackup');
     return (
       <Wrapper>
         <LeftSection>
-          <AccountsOverview accounts={accounts} currentAccountIndex={currentAccountIndex} switchAccount={setCurrentAccount} />
+          <AccountsOverview walletName={displayName} accounts={accounts} currentAccountIndex={currentAccountIndex} switchAccount={setCurrentAccount} />
           {!hasBackup && (
             <BackupReminder onClick={this.navigateToBackup}>
               <FullCrossIcon src={leftSideTIcon} />
@@ -108,7 +104,7 @@ class Wallet extends Component<Props, State> {
         </LeftSection>
         <CorneredWrapper>
           <RightSection>
-            <HorizontalBar src={smallHorizontalSideBar} />
+            <SmallHorizontalPanel />
             <Switch>
               {routes.wallet.map((route) => (
                 <Route exact key={route.path} path={route.path} component={route.component} />
@@ -121,20 +117,25 @@ class Wallet extends Component<Props, State> {
     );
   }
 
-  componentDidMount() {
-    // this.getBalance();
+  async componentDidMount() {
+    const { isConnected, getBalance } = this.props;
+    if (isConnected) {
+      try {
+        await getBalance();
+        this.getBalanceInterval = setInterval(async () => {
+          await getBalance();
+        }, 60000);
+      } catch (error) {
+        this.setState(() => {
+          throw error;
+        });
+      }
+    }
   }
 
-  getBalance = async () => {
-    const { getBalance } = this.props;
-    try {
-      await getBalance();
-    } catch (error) {
-      this.setState(() => {
-        throw error;
-      });
-    }
-  };
+  componentWillUnmount() {
+    this.getBalanceInterval && clearInterval(this.getBalanceInterval);
+  }
 
   navigateToBackup = () => {
     const { history } = this.props;
@@ -143,6 +144,8 @@ class Wallet extends Component<Props, State> {
 }
 
 const mapStateToProps = (state) => ({
+  isConnected: state.node.isConnected,
+  displayName: state.wallet.meta.displayName,
   accounts: state.wallet.accounts,
   currentAccountIndex: state.wallet.currentAccountIndex
 });
