@@ -26,9 +26,18 @@ const ipcConsts = {
   SET_NODE_IP_FAILURE: 'SET_NODE_IP_FAILURE',
   NOTIFICATION_CLICK: 'NOTIFICATION_CLICK',
   DELETE_FILE: 'DELETE_FILE',
+  WIPE_OUT: 'WIPE_OUT',
   TOGGLE_AUTO_START: 'TOGGLE_AUTO_START',
   IS_AUTO_START_ENABLED_REQUEST: 'IS_AUTO_START_ENABLED_REQUEST',
   IS_AUTO_START_ENABLED_REQUEST_RESPONSE: 'IS_AUTO_START_ENABLED_REQUEST_RESPONSE',
+  START_NODE: 'START_NODE',
+  START_NODE_SUCCESS: 'START_NODE_SUCCESS',
+  START_NODE_FAILURE: 'START_NODE_FAILURE',
+  HARD_REFRESH: 'HARD_REFRESH',
+  REQUEST_CLOSE: 'REQUEST_CLOSE',
+  QUIT_NODE: 'QUIT_NODE',
+  QUIT_APP: 'QUIT_APP',
+  KEEP_RUNNING_IN_BACKGROUND: 'KEEP_RUNNING_IN_BACKGROUND',
   // gRPC calls
   // Node management
   CHECK_NODE_CONNECTION: 'CHECK_NODE_CONNECTION',
@@ -43,9 +52,9 @@ const ipcConsts = {
   GET_GENESIS_TIME: 'GET_GENESIS_TIME',
   GET_GENESIS_TIME_SUCCESS: 'GET_GENESIS_TIME_SUCCESS',
   GET_GENESIS_TIME_FAILURE: 'GET_GENESIS_TIME_FAILURE',
-  GET_UPCOMING_REWARDS: 'GET_UPCOMING_REWARDS',
-  GET_UPCOMING_REWARDS_SUCCESS: 'GET_UPCOMING_REWARDS_SUCCESS',
-  GET_UPCOMING_REWARDS_FAILURE: 'GET_UPCOMING_REWARDS_FAILURE',
+  GET_UPCOMING_AWARDS: 'GET_UPCOMING_AWARDS',
+  GET_UPCOMING_AWARDS_SUCCESS: 'GET_UPCOMING_AWARDS_SUCCESS',
+  GET_UPCOMING_AWARDS_FAILURE: 'GET_UPCOMING_AWARDS_FAILURE',
   SET_AWARDS_ADDRESS: 'SET_AWARDS_ADDRESS',
   SET_AWARDS_ADDRESS_SUCCESS: 'SET_AWARDS_ADDRESS_SUCCESS',
   SET_AWARDS_ADDRESS_FAILURE: 'SET_AWARDS_ADDRESS_FAILURE',
