@@ -1,9 +1,9 @@
 const nodeConsts = {
-  COMMITMENT_SIZE: 200, // GB
-  DRIVE_SPACE_BUFFER: 50, // GB
+  COMMITMENT_SIZE: 256, // GB
   NOT_MINING: 1,
   IN_SETUP: 2,
-  IS_MINING: 3
+  IS_MINING: 3,
+  TIME_BETWEEN_LAYERS: 300000
 };
 
 export default nodeConsts;
